refactor(sybra): extract LifecycleManager and services registry
- Add lifecycle.go: LifecycleManager with StartManagers, StartPollers,
  StartWatchers phases; moves 7 background-service methods out of app.go
- Add services.go: wireServices, ServiceRegistry, BindTargets consolidated
  as the explicit service ownership registry
- app.go: dep graph comment, Startup delegates to lm.Start* phases,
  extract initFileWatcher, fix duplicate startConfigWatcher call, -302 lines
